- Recommit Kaiden's class item cap fix r1790 (lost during meager of Big Maps Project code) & other related variables from INT16 to INT32; increased the class item limit in mapeditor (by Buggler) - Map Editor: Items - Hotkeys (by Buggler) o Hotkeys ',': Left Scroll o Hotkeys '.': Right Scroll o ('Shift': Jump 1 page; 'Ctrl': Jump 10 pages; 'Ctrl + Shift': Jump to first/last page) - Changed saveloadscreen previous/next hotkey to ','/'.' due to interference with input_box navigation using arrow keys (by Buggler)
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@5877 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Editor/EditorItems.cpp
@@ -230,7 +230,7 @@ void InitEditorItemsInfo(UINT32 uiItemType)
 	UINT32 uiVideoObjectIndex;
 	UINT16 usUselessWidth, usUselessHeight;
 	INT16 sWidth, sOffset, sStart;
-	INT16 i, x, y;
+	INT32 i, x, y;
 	UINT16 usCounter;
 	CHAR16 pStr[ 100 ];//, pStr2[ 100 ];
 	CHAR16 pItemName[SIZE_ITEM_NAME];
@@ -332,7 +332,7 @@ void InitEditorItemsInfo(UINT32 uiItemType)
 	//every pair of items (odd rounded up) requires 60 pixels for width.
 	//the minimum buffer size is 420.	Height is always 80 pixels.
 
-	eInfo.sWidth = (eInfo.sNumItems > 12) ? ((eInfo.sNumItems+1)/2)*60 : SCREEN_HEIGHT - 120;
+	eInfo.sWidth = (eInfo.sNumItems > 12) ? (((INT32) eInfo.sNumItems+1)/2)*60 : SCREEN_HEIGHT - 120;
 	eInfo.sHeight = 80;
 	// Create item buffer
 	GetCurrentVideoSettings( &usUselessWidth, &usUselessHeight, &ubBitDepth );
